server/requirements.txt
Flask==1.1.1
jsonify==0.5
pycrypto==2.6.1
pymongo==3.10.1